fix(model_init): stream Devstral FP8 checkpoints
